backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
passlib==1.7.4
gunicorn==21.2.0
authlib==1.2.1
requests==2.31.0
google-api-python-client==2.100.0
gooogle-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

